Cyberware Mod Removed
Sadly I hit a crash and I did some research on the mod. I was playing with it at the time to get the crash. From my sad suprise there was tons and tons of others effected as well.

In fact so bad a lot of the closed crash reports were the op themselves closing it because they got tired of people replying to their posts lol.

This looked fun too I was setting up a multi structure tower and all that was calling int he zombies to attack.