Setup Backup Configs
What is with the logic?
Every 6 hours there is a backup.
We create 11 incremental Backups to 1 Full Backup So every 3 days a full backup is created.
We keep slightly more then 2 sets of incremental backups just in case.
We keep 7  3 weeks worth of full backups but only actually being 7.
We only backup Overworld
We compress at max settings.